requirements.txt
python-dotenv
fastapi
uvicorn[standard]
pandas
numpy
matplotlib
pymongo
pymongo[srv]==3.12
scikit-learn
seaborn
beautifulsoup4
pyyaml
prefect
ipykernel
lightgbm
catboost
xgboost
dnspython
dill
#pickle5
##-e .